test: change tests to use jest
