+++ sites/example-project/jest.config.js
@@ -1,0 +1,6 @@
+/** @type {import('jest').Config} */
+const config = {
+	testMatch: ['**/*.spec.js']
+};
+
+export default config
