+++ sites/example-project/package.json
@@ -38,16 +38,14 @@
 	},
 	"devDependencies": {
 		"@sveltejs/package": "1.0.1",
-		"jest": "^28.1.3",
-		"vite": "^4.0.4",
 		"@sveltejs/vite-plugin-svelte": "^2.0.3",
+		"@types/jest": "^29.5.0",
+		"autoprefixer": "^10.4.7",
+		"jest": "^28.1.3",
 		"postcss": "^8.4.14",
 		"postcss-load-config": "^4.0.1",
 		"svelte-preprocess": "^4.10.7",
-		"autoprefixer": "^10.4.7",
-		"tailwindcss": "^3.1.5"
-	},
-	"jest": {
-		"verbose": true
+		"tailwindcss": "^3.1.5",
+		"vite": "^4.0.4"
 	}
 }
